[Refactor] Extract createNetworkedEntityFromRemote
